JasazEyesV1/requirements.txt
discord.py>=2.3.2
asyncio
aiohttp